mcp: carve events.rs (pure move)
The AppEvent observation layer leaves mod.rs: spawn_event_listener,
apply_observed_event_to_mcp_state, the session-capability and
context-snapshot appliers, persisted-log session hydration, and
spawn_http_observation_listener — 1,515 body lines and 27 tests (2,044
lines) verbatim. Fixtures test_state/TEST_ENV_LOCK widened pub(crate)
in mod.rs's tests for the moved referents.

Same glue as the prior carves; purity-checked; mcp:: tests 178/178.
